+++ src/components/MapPin.jsx
@@ -13,6 +13,9 @@ import BookIcon from '@mui/icons-material/Book';
 import CreateIcon from '@mui/icons-material/Create';
 import LocalLibraryIcon from '@mui/icons-material/LocalLibrary';
 
+// Theme helpers
+import useMediaQuery from '@mui/material/useMediaQuery';
+
 const PinBox = styled(Box)(({ theme }) => ({
   display: "flex",
   position: "relative",
@@ -26,6 +29,8 @@ const PinBox = styled(Box)(({ theme }) => ({
 export default function MapPin({data, onClickPin}) {
 
   const theme = useTheme();
+  const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
+
   const [icon, setIcon] = useState("");
   const [color, setColor] = useState("pinDefault");
   const [isDragging, setIsDragging] = useState(false);
@@ -98,23 +103,31 @@ export default function MapPin({data, onClickPin}) {
     setIcon(icon);
   }, [data.id]);
 
-  // We perform this trio of event handling to avoid triggering the popups
+  // We perform this set of event handling to avoid triggering the popups
   // when the user is dragging the map to pan around.
 
+  // Mobile only
+
   const handleMouseDown = () => {
-    setIsDragging(false);
+    isMobile && setIsDragging(false);
   };
 
   const handleMouseMove = () => {
-    setIsDragging(true);
+    isMobile && setIsDragging(true);
   };
 
   const handleMouseUp = () => {
-    if (!isDragging) {
+    if (!isDragging && isMobile) {
       onClickPin(data);
     }
     setIsDragging(false);
   };
+
+  // Desktop only
+
+  const handleClick = () => {
+    !isMobile && onClickPin(data);
+  }
 
   return (
     <PinBox
@@ -122,6 +135,7 @@ export default function MapPin({data, onClickPin}) {
       onMouseDown={handleMouseDown}
       onMouseMove={handleMouseMove}
       onMouseUp={handleMouseUp}
+      onClick={handleClick}
     >
       <IconBox className="sff-map-icon">{icon}</IconBox>
       {data.showLabel && (