+++ CHANGELOG.md
@@ -8,6 +8,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Map:
   - Fix: Performance improvements
 
+Sidebar:
+  - Fix: Many performance improvements
+
+Pins:
+  - Improve: On desktop, pins sometimes hard to click.
+
 ## [2.0.0] - 2024-05-16 (beta)
 
 General: